Refactor balances adjustment to move network areas in dedicated package (#238)
Move network areas interface and implementation in a dedicated module.
Refactor tests logging to limit to error messages.
Add documentation for Network Area concept, definitions, and calculation methods
