+++ balances-adjustment/src/test/java/com/powsybl/balances_adjustment/balance_computation/extension/ExtendedBalanceComputationTest.java
@@ -7,7 +7,7 @@
 package com.powsybl.balances_adjustment.balance_computation.extension;
 
 import com.powsybl.balances_adjustment.balance_computation.*;
-import com.powsybl.balances_adjustment.util.CountryAreaFactory;
+import com.powsybl.network_area.CountryAreaFactory;
 import com.powsybl.computation.ComputationManager;
 import com.powsybl.computation.local.LocalComputationManager;
 import com.powsybl.iidm.modification.scalable.Scalable;
