+++ cne/cne-converter/pom.xml
@@ -77,8 +77,8 @@
             <scope>test</scope>
         </dependency>
          <dependency>
-             <groupId>org.slf4j</groupId>
-             <artifactId>slf4j-simple</artifactId>
+             <groupId>ch.qos.logback</groupId>
+             <artifactId>logback-classic</artifactId>
              <scope>test</scope>
          </dependency>
          <dependency>
