+++ docs/index.rst
@@ -5,4 +5,5 @@ PowSyBl ENTSO-E's user documentation
    :maxdepth: 2
 
    flow_decomposition/index
+   network_area/index.md
    glsk/glsk.md
