+++ emf/src/test/resources/logback-test.xml
@@ -14,11 +14,9 @@
             <pattern>%-5p %-20C{1} | %m%n</pattern>
         </encoder>
     </appender>
-    <root level="info">
+    <root level="error">
         <appender-ref ref="STDOUT"/>
     </root>
-    <logger name="com.bigdata" level="error"/>
-    <logger name="com.powsybl.triplestore.TripleStore" level="info"/>
-    <logger name="com.powsybl.cgmes.triplestore.CgmesModelTripleStore" level="info"/>
-    <logger name="com.powsybl.cgmes.conversion" level="error"/>
+    <!-- To be modified if debug analysis needed -->
+    <logger name="com.powsybl.emf" level="error"/>
 </configuration>
