--- glsk/glsk-commons/src/test/resources/TestCase12Nodes.uct
@@ -1,38 +1,0 @@
-##C 2007.05.01
-##N
-##ZBE
-BBE1AA1  BE1          0 2 400.00 2500.00 0.00000 -1500.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-BBE2AA1  BE2          0 2 400.00 1000.00 0.00000 -3000.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-BBE3AA1  BE3          0 2 400.00 1500.00 0.00000 -2500.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-##ZDE
-DDE1AA1  DE1          0 2 400.00 3500.00 0.00000 -2500.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-DDE2AA1  DE2          0 2 400.00 3000.00 0.00000 -2000.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-DDE3AA1  DE3          0 2 400.00 2000.00 0.00000 -1500.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-##ZFR
-FFR1AA1  FR1          0 2 400.00 1000.00 0.00000 -2000.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-FFR2AA1  FR2          0 2 400.00 3500.00 0.00000 -2000.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-FFR3AA1  FR3          0 2 400.00 1500.00 0.00000 -3000.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-##ZNL
-NNL1AA1  NL1          0 2 400.00 1000.00 0.00000 -1500.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-NNL2AA1  NL2          0 2 400.00 1000.00 0.00000 -500.00 0.00000 9000.00 -9000.0 9000.00 -9000.0
-NNL3AA1  NL3          0 2 400.00 2500.00 0.00000 -2000.0 0.00000 9000.00 -9000.0 9000.00 -9000.0
-##L
-BBE1AA1  BBE2AA1  1 0 0.0000 10.000 0.000000   5000
-BBE1AA1  BBE3AA1  1 0 0.0000 10.000 0.000000   5000
-FFR1AA1  FFR2AA1  1 0 0.0000 10.000 0.000000   5000
-FFR1AA1  FFR3AA1  1 0 0.0000 10.000 0.000000   5000
-FFR2AA1  FFR3AA1  1 0 0.0000 10.000 0.000000   5000
-DDE1AA1  DDE2AA1  1 0 0.0000 10.000 0.000000   5000
-DDE1AA1  DDE3AA1  1 0 0.0000 10.000 0.000000   5000
-DDE2AA1  DDE3AA1  1 0 0.0000 10.000 0.000000   5000
-NNL1AA1  NNL2AA1  1 0 0.0000 10.000 0.000000   5000
-NNL1AA1  NNL3AA1  1 0 0.0000 10.000 0.000000   5000
-NNL2AA1  NNL3AA1  1 0 0.0000 10.000 0.000000   5000
-FFR2AA1  DDE3AA1  1 0 0.0000 10.000 0.000000   5000
-DDE2AA1  NNL3AA1  1 0 0.0000 10.000 0.000000   5000
-NNL2AA1  BBE3AA1  1 0 0.0000 10.000 0.000000   5000
-BBE2AA1  FFR3AA1  1 0 0.0000 10.000 0.000000   5000
-##T
-BBE2AA1  BBE3AA1  1 0 400.0 400.0 1000. 0.0000 10.000 0.000000 0.0	     5000 PST
-##R
-BBE2AA1  BBE3AA1  1                    -0.68 90.00 16  0        SYMM
