+++ glsk/glsk-commons/src/test/resources/logback-test.xml
@@ -1,11 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (c) 2026, RTE (http://www.rte-france.com)
+    This Source Code Form is subject to the terms of the Mozilla Public
+    License, v. 2.0. If a copy of the MPL was not distributed with this
+    file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+-->
 <configuration>
     <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
         <encoder>
-            <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</pattern>
+            <!-- <pattern>%-5p %d{HH:mm:ss.SSS} %-20C{1} | %m%n</pattern> -->
+            <pattern>%-5p %-20C{1} | %m%n</pattern>
         </encoder>
     </appender>
-
-    <root level="debug">
-        <appender-ref ref="STDOUT" />
+    <root level="error">
+        <appender-ref ref="STDOUT"/>
     </root>
+    <!-- To be modified if debug analysis needed -->
+    <logger name="com.powsybl.glsk.commons" level="error"/>
 </configuration>
