+++ network-area/src/test/java/com/powsybl/network_area/BorderBasedVoltageLevelsAreaTest.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.balances_adjustment.util;
+package com.powsybl.network_area;
 
 import com.powsybl.iidm.network.Network;
 import org.junit.jupiter.api.BeforeEach;
