+++ network-area/src/test/java/com/powsybl/network_area/InjectionBasedVoltageLevelsAreaTest.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.balances_adjustment.util;
+package com.powsybl.network_area;
 
 import com.powsybl.iidm.network.Network;
 import org.junit.jupiter.api.BeforeEach;
