+++ network-area/src/test/resources/controlArea.xiidm
