+++ network-area/src/test/resources/testCaseNanInNetPositionComputation.uct
