+++ network-area/src/test/resources/testCaseSpecialDevices.xiidm
